--- src/features/auth/internal/AuthContext/index.ts
@@ -1,4 +1,0 @@
-export { useAuth } from './internal/hooks/useAuth';
-export { useGuard } from './internal/hooks/useGuard';
-export type { UseGuardProps } from './internal/hooks/useGuard';
-export { default as AuthProvider } from './internal/AuthProvider.component';
