--- src/features/auth/internal/AuthContext/internal/AuthContext.ts
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-import { AuthContextType } from './AuthContext.types';
-
-export const AuthContext = createContext<AuthContextType | undefined>(
-  undefined
-);
